Move examples into own directory
