+++ docs/00-toc.md
@@ -20,7 +20,7 @@ Apps are lighweight interactive add-ons to mattermost. Apps can:
 - [Anatomy](01-anatomy-hello.md) of a simple app.
 
 #### Functions, Calls
-- Example: [Hello, World](/server/examples/go/helloworld/hello.go#L45) `send`
+- Example: [Hello, World](/examples/go/helloworld/hello.go#L45) `send`
   function.
 - [Post Menu Message Flow](02-example-post-menu.md) - message flow to define a Post Menu
   action, then have a user click it.
@@ -38,7 +38,7 @@ Apps are lighweight interactive add-ons to mattermost. Apps can:
   controls context expansion.
 
 #### Forms
-- Example: [Hello, World](/server/examples/go/helloworld/send_form.json) `send` form.
+- Example: [Hello, World](/examples/go/helloworld/send_form.json) `send` form.
 - [Interactive Message Flows](03-example-interactivity.md) - message flow to define a
   Post Menu action, then have a user click it.
 - godoc [Form](https://pkg.go.dev/github.com/mattermost/mattermost-plugin-apps/apps#Form) - [local](http://localhost:6060/pkg/github.com/mattermost/mattermost-plugin-apps/apps#Form)
@@ -46,14 +46,14 @@ Apps are lighweight interactive add-ons to mattermost. Apps can:
 
 
 #### Bindings and Locations
-- Example: [Hello, World](/server/examples/go/helloworld/bindings.json) bindings.
+- Example: [Hello, World](/examples/go/helloworld/bindings.json) bindings.
 - godoc [Binding](https://pkg.go.dev/github.com/mattermost/mattermost-plugin-apps/apps#Binding) - [local](http://localhost:6060/pkg/github.com/mattermost/mattermost-plugin-apps/apps#Binding)
 
 #### Autocomplete
-- Example: [Hello, World](/server/examples/go/helloworld/bindings.json) bindings.
+- Example: [Hello, World](/examples/go/helloworld/bindings.json) bindings.
 
 #### Modals
-- Example: [Hello, World](/server/examples/go/helloworld/send_form.json) `send` form.
+- Example: [Hello, World](/examples/go/helloworld/send_form.json) `send` form.
 
 #### In-Post Interactivity
