+++ examples/go/hello-oauth2/bindings.json
