+++ examples/go/hello-oauth2/icon.png
