+++ examples/go/hello-oauth2/manifest.json
@@ -4,7 +4,7 @@
 	"display_name": "Hello, OAuth2!",
 	"app_type": "http",
 	"root_url": "http://localhost:8080",
-	"homepage_url": "https://github.com/mattermost/mattermost-plugin-apps/server/examples/go/hello-oauth2",
+	"homepage_url": "https://github.com/mattermost/mattermost-plugin-apps/examples/go/hello-oauth2",
 	"requested_permissions": [
 		"act_as_admin",
 		"act_as_user",
