+++ examples/go/hello-webhooks/bindings.json
