+++ examples/go/hello-webhooks/icon.png
