+++ examples/go/hello-webhooks/manifest.json
@@ -4,7 +4,7 @@
 	"display_name": "Hello, Webhooks!",
 	"app_type": "http",
 	"root_url": "http://localhost:8080",
-	"homepage_url": "https://github.com/mattermost/mattermost-plugin-apps/server/examples/go/hello-webhooks",
+	"homepage_url": "https://github.com/mattermost/mattermost-plugin-apps/examples/go/hello-webhooks",
 	"requested_permissions": [
 		"act_as_admin",
 		"act_as_bot",
