+++ examples/go/hello/bindings.go
