+++ examples/go/hello/http_hello/functions.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/mattermost/mattermost-plugin-apps/apps"
-	"github.com/mattermost/mattermost-plugin-apps/server/examples/go/hello"
+	"github.com/mattermost/mattermost-plugin-apps/examples/go/hello"
 	"github.com/mattermost/mattermost-plugin-apps/server/utils/httputils"
 	"github.com/pkg/errors"
 )
