+++ examples/go/hello/install.go
