+++ examples/go/hello/send_survey.go
