+++ examples/go/hello/submit_survey.go
