+++ examples/go/hello/survey.go
