+++ examples/go/hello/user_joined_channel.go
