+++ examples/go/helloworld/bindings.json
