+++ examples/go/helloworld/manifest.json
