+++ examples/js/aws_hello/hello_app/manifest.json
