+++ examples/js/aws_hello/hello_app/on_activate/index.js
