+++ examples/js/aws_hello/helloapp.go
